requirements.txt
flask
pyserial
psycopg2-binary